Correct bcdDFUVersion and wDetachTimeOut in DFU functional descriptor
The firmware matches pretty much the DFU 1.1 specification.

This is a follow up of 95e1150ddcbb6e0e4c4e2ff9f85a0e6d9d620407
which only changed the separately defined u8_usbFunctionalDescriptor
but not the one included in the configuration descriptor. And which
also got it wrong (0.1).

(Ideally the code should be cleaned up to only specify this in one place)

Also fix the detach timeout which was set to 64 seconds. I believe
the intended value was 255 ms like stated in the comment. Since the
Maple devices are not detached (transition between run-time mode and
DFU mode, issued over the DFU interface), this is not used anyway.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>
Acked-by: Marti Bolivar <mbolivar@leaflabs.com>