Add DFU_UPLOAD support
The DFU_UPLOAD support allows copying the contents of the flash or RAM
memory of the Maple to the computer.

Note that we use the wValue (BlockNum in the DFU standard) to calculate
the target address. We therefore rely on the host program (for instance
dfu-util) to set (increase) wValue for each block.

This use of wValue is not mandated by the DFU standard, which only says
wValue should be incremented by the host "to provide useful context" to
the device, but this implementation allows random access of memory
locations.

On the other hand the DFU_DNLOAD code does not make use of BlockNum but
simply increases its memory pointer for each block received. Maybe this
was done because reading out wValue was broken due to endian mix-up?

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>