Remove spurious configuration of whole lower GPIOA bank
This is not much of an issue as long as SetupLED() is the first place
the GPIOA bank is configured and the ports are inputs anyway, but it
would cause trouble if something tried to configure any of the ports
before SetupLED().

Note that ports are by default floating inputs after reset,
but this configuration was setting them to analog inputs instead.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>